Move tests to agent/dns/dns_test.go
