+++ .changelog/10611.txt
@@ -1,3 +1,3 @@
-```release-note:feature
+```release-note:improvement
 config: add `dns_config.recursor_strategy` flag to control the order which DNS recursors are queried
 ```
